[REF] hr_cae: extract hr_employee in own file
