+++ hr_cae/__manifest__.py
@@ -11,7 +11,12 @@
     "version": "12.0.1.0.0",
     "license": "AGPL-3",
     "depends": ["hr", "l10n_fr_department", "hr.recruitment"],
-    "data": ["security/ir.model.access.csv", "views/hr.xml", "data/data.xml"],
+    "data": [
+        "security/ir.model.access.csv",
+        "views/hr.xml",
+        "hr_employee.xml",
+        "data/data.xml",
+    ],
     "demo": [],
     "installable": True,
     "application": False,
