+++ hr_cae/models/__init__.py
@@ -1,1 +1,2 @@
 from . import hr
+from . import hr_employee
