+++ hr_cae/views/hr.xml
@@ -4,17 +4,6 @@
      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 <odoo>
-    <record id="view_employee_form" model="ir.ui.view">
-        <field name="name">hr.employee.form</field>
-        <field name="model">hr.employee</field>
-        <field name="inherit_id" ref="hr.view_employee_form"/>
-        <field name="arch" type="xml">
-            <field name="address_id" position="after">
-                <field name="country_department_id"/>
-            </field>
-        </field>
-    </record>
-
     <record id="open_view_origin_status_form" model="ir.actions.act_window">
         <field name="name">Origin Status</field>
         <field name="res_model">hr.origin.status</field>
